+++ flogin/flow/plugin_metadata.py
@@ -6,7 +6,7 @@ from typing import TYPE_CHECKING, Any, Awaitable, Literal
 from .base import Base, add_prop
 
 if TYPE_CHECKING:
-    from .client import FlowLauncherAPI
+    from .api import FlowLauncherAPI
 
 __all__ = ("PluginMetadata",)
 
@@ -82,7 +82,7 @@ class PluginMetadata(Base):
 
         Registers a new keyword with flow for the plugin.
 
-        This is a shortcut to :func:`~flogin.flow_api.client.FlowLauncherAPI.add_keyword`
+        This is a shortcut to :func:`~flogin.flow.api.FlowLauncherAPI.add_keyword`
 
         Parameters
         --------
@@ -101,7 +101,7 @@ class PluginMetadata(Base):
 
         Removes a keyword from the plugin.
 
-        This is a shortcut to :func:`~flogin.flow_api.client.FlowLauncherAPI.remove_keyword`
+        This is a shortcut to :func:`~flogin.flow.api.FlowLauncherAPI.remove_keyword`
 
         Parameters
         --------