+++ docs/whats_new.rst
@@ -14,6 +14,8 @@ Breaking Changes
     - Remove ``flogin.errors.SettingNotFound``
 - Rewrite the CLI commands
 - Remove ``flogin.conditions.MultiCondition`` in favor of :class:`flogin.conditions.AnyCondition` and :class:`flogin.conditions.AllCondition`
+- Rename the ``flogin.flow_api`` directory to ``flogin.flow``
+    - Rename ``flogin.flow_api.client.py`` to ``flogin.flow.api.py``
 
 New Features
 ~~~~~~~~~~~~
@@ -31,11 +33,27 @@ New Features
 - Make :attr:`flogin.jsonrpc.results.Result.title` optional
 - Add :class:`flogin.conditions.AnyCondition`
 - Add :class:`flogin.conditions.AllCondition`
+- Add ``flogin.flow.settings.py``
+    - Add :class:`flogin.flow.settings.CustomExplorerViewModel`
+    - Add :class:`flogin.flow.settings.CustomBrowserViewModel`
+    - Add :class:`flogin.flow.settings.CustomPluginHotkey`
+    - Add :class:`flogin.flow.settings.CustomShortcutModel`
+    - Add :class:`flogin.flow.settings.HttpProxy`
+    - Add :class:`flogin.flow.settings.PartialPlugin`
+    - Add :class:`flogin.flow.settings.PluginsSettings`
+    - Add :class:`flogin.flow.settings.FlowSettings`
+- Add ``flogin.flow.enums.py``
+    - Add :class:`flogin.flow.enums.LastQueryMode`
+    - Add :class:`flogin.flow.enums.ColorSchemes`
+    - Add :class:`flogin.flow.enums.SearchWindowScreens`
+    - Add :class:`flogin.flow.enums.SearchWindowAligns`
+    - Add :class:`flogin.flow.enums.AnimationSpeeds`
+    - Add :class:`flogin.flow.enums.SearchPrecisionScore`
 
 Bug Fixes
 ~~~~~~~~~
 
-- Fix bug where :func:`flogin.flow_api.client.FlowLauncherAPI.update_results` does not register the results, so callbacks do not get triggered.
+- Fix bug where :func:`flogin.flow.api.FlowLauncherAPI.update_results` does not register the results, so callbacks do not get triggered.
 - Fix typing bug with :func:`flogin.plugin.Plugin.register_search_handlers` and :func:`flogin.plugin.Plugin.register_search_handler` due to :class:`flogin.search_handler.SearchHandler` being a generic.
 - Fix bug where ``Glyph`` was not included in ``ResultConstructorArgs``
 - Fix bug with the ``PluginT`` TypeVar not being marked as covariant
@@ -73,7 +91,7 @@ Bug Fixes
 - Fixed bug with the `create plugin.json <cli-create-plugin-json>`_ CLI command dumping the wrong data.
 - Add the missing dependency `click <https://pypi.org/project/click/>`_
 - Fix bug where :attr:`flogin.search_handler.SearchHandler.plugin` is ``None`` when :obj:`flogin.search_handler.SearchHandler.condition` is ran.
-- Fix an ``AttributeError`` that gets raised in a couple of :class:`~flogin.flow_api.client.FlowLauncherAPI` methods
+- Fix an ``AttributeError`` that gets raised in a couple of :class:`~flogin.flow.api.FlowLauncherAPI` methods
 - Log fatal errors that cause startup crashes to prevent them from being hidden.
 - Fix bug where ``rounded_icon`` was not present in ``ResultConstructorArgs``
 - Fix a ``NameError`` that gets raised in the default :func:`flogin.jsonrpc.results.Result.callback`